requirements.txt
absl-py==1.0.0
astor==0.8.1
bchlib==0.14.0
blessings==1.7
cached-property==1.5.2
certifi==2016.9.26
cloudpickle==2.0.0
cycler==0.11.0
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.43.0
h5py==3.1.0
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
numpy==1.18.5
nvidia-cublas==11.2.1.74
nvidia-cuda-cupti==11.1.69
nvidia-cuda-nvcc==11.1.74
nvidia-cuda-nvrtc==11.1.105
nvidia-cuda-runtime==11.1.74
nvidia-cudnn==8.0.4.30
nvidia-cufft==10.3.0.74
nvidia-curand==10.2.2.74
nvidia-cusolver==11.0.0.74
nvidia-cusparse==11.2.0.275
nvidia-dali-cuda110==0.27.0
nvidia-dali-nvtf-plugin==0.27.0+nv20.11
nvidia-horovod==0.20.2+nv20.11
nvidia-ml-py3==7.352.0
nvidia-nccl==2.8.2
nvidia-pyindex==1.0.9
nvidia-tensorboard==1.15.0+nv20.11
nvidia-tensorflow==1.15.4+nv20.11
nvidia-tensorrt==7.2.1.6
opencv-python==4.5.5.62
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.19.3
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
reedsolo==1.5.4
scipy==1.5.2
screen==1.0.1
six==1.16.0
stn==1.0.1
tensorboard @ https://pypi.ngc.nvidia.com/tensorboard/tensorboard-1.15.0-py2.py3-none-any.whl
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing_extensions==4.0.1
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
